7 authorization

+++ app/controllers/topics_controller.rb
@@ -1,10 +1,12 @@
 class TopicsController < ApplicationController
+  after_action :verify_authorized, only: :destroy
+
   def index
     @topics = Topic.all
   end
 
   def show
-    @topic = Topic.find(params[:id])
+    @topic = Topic.friendly.find(params[:id])
   end
 
   def new
@@ -12,7 +14,7 @@ class TopicsController < ApplicationController
   end
 
   def create
-    @topic = current_user.topics.new(topic_params)
+    @topic = Topic.new(topic_params)
 
     if @topic.save
       flash[:notice] = "Topic was saved successfully."
@@ -24,11 +26,11 @@ class TopicsController < ApplicationController
   end
 
   def edit
-    @topic = Topic.find(params[:id])
+    @topic = Topic.friendly.find(params[:id])
   end
 
   def update
-    @topic = Topic.find(params[:id])
+    @topic = Topic.friendly.find(params[:id])
     @topic.assign_attributes(topic_params)
 
     if @topic.save
@@ -41,7 +43,8 @@ class TopicsController < ApplicationController
   end
 
   def destroy
-    @topic = current_user.topics.find(params[:id])
+    @topic = Topic.friendly.find(params[:id])
+    authorize @topic 
 
     if @topic.destroy
       flash[:notice] = "Topic was successfully deleted."
